cambios metodo eliminar

+++ project/views.py
@@ -285,12 +285,13 @@ import simplejson
 @login_required(login_url="/login-profesor")
 @user_passes_test(restringir_estudiante, login_url='/login-profesor')
 def eliminar_estudiante(request):
-
-    pk = request.POST.get('estudiante_id')
-    estudiante = User.objects.get(pk=pk)
-    estudiante.delete()
-    response = {}
-    return JsonResponse(response)
+    
+    if request.method == 'POST':
+        pk = request.POST.get('estudiante_id')
+        estudiante = User.objects.get(pk=pk)
+        estudiante.delete()
+        response = {}
+        return JsonResponse(response)
 
 import xhtml2pdf.pisa as pisa
 from StringIO import StringIO
